+++ package.json
@@ -13,6 +13,7 @@
     "axios": "^1.2.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-query": "^3.39.2",
     "react-scripts": "5.0.0",
     "typescript": "^4.5.5"
   },
