+++ src/components/Exchange.tsx
@@ -1,0 +1,12 @@
+import ExchangeInfo from "./ExchangeInfo";
+import ExchangeInput from "./ExchangeInput";
+
+export default function Exchange() {
+  return (
+    <div className="App">
+      <ExchangeInfo />
+      <hr />
+      <ExchangeInput />
+    </div>
+  );
+}
